Add: Pino dependency
This commit adds Pino as a dependency so we can do logging. It outputs
to both standard output and a file based on the settings passed through
environment variables.

Also added is the gracefulShutdown() function. This will run whenever
the program terminates and is meant to shutdown/disconnect all services
cleanly before appliction termination.